% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert a dataframe to a matrix}
\usage{
df_to_matrix(x)
}
\arguments{
\item{x}{A dataframe.}
}
\description{
This function converts a dataframe to a matrix, with the first column of the dataframe becoming the matrix rownames.
}
\examples{
df_to_matrix(my.dataframe)
}
