% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_divergences.R
\name{get_stan_divergences}
\alias{get_stan_divergences}
\title{Summarize divergent iterations in a Stan model object}
\usage{
get_stan_divergences(fit.stan.model)
}
\arguments{
\item{fit.stan.model}{A Stan model object.}
}
\description{
This function summarizes the number of divergent iterations observed across all Markov chains in a fit Stan model object.
}
\examples{
get_stan_divergences(my.stan.model)
}
