% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilogit.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit function}
\usage{
ilogit(x)
}
\arguments{
\item{x}{A number.}
}
\description{
This function performs the inverse logit transformation.
}
\examples{
ilogit(my.number)
}
