% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_NA_cols.R
\name{remove_NA_cols}
\alias{remove_NA_cols}
\title{Remove columns from an array if they contain all NA values}
\usage{
remove_NA_cols(x)
}
\arguments{
\item{x}{A dataframe or other array.}
}
\description{
A convenience function to clean up your data by removing any columns that will not be useful since they consist entirely of NA values.
}
\examples{
remove_NA_cols(my.dataframe)
}
