% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reskew_binom_test.R
\name{reskew_binom_test}
\alias{reskew_binom_test}
\title{Perform the exact binomial test with custom output format}
\usage{
reskew_binom_test(numeric.vector)
}
\arguments{
\item{numeric.vector}{A binary numeric vector.}
}
\description{
This function is a wrapper for \code{binom.test()} with a custom output format. Specifically, the function returns a dataframe with columns for the mean of the input vector as well as min and max values calculated from \code{binom.test}. This function is particularly useful in conjunction with the \code{stat_summary()} function in \pkg{ggplot2} for plotting summary statistics of binary data.
}
\examples{
test.vector <- rep(c(1, 0), 25)

reskew_binom_test(test.vector)
}
