% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reskew_precis_plot.R
\name{reskew_precis_plot}
\alias{reskew_precis_plot}
\title{A custom precis plot function}
\usage{
reskew_precis_plot(x, y, pars, labels, cex, col.ci = "black",
  xlab = "Value", xlim...)
}
\description{
This function is a modified version of \code{rethinking}'s plot method for \code{precis} objects. It allows for custom x-axis limits and specification of row labels.
}
\examples{
my.precis.object <- rethinking::precis(my.dataframe, prob = 0.95)

reskew_precis_plot(my.precis.object, xlab = "Parameter Estimate",
                   labels = my.custom.labels, xlim = c(-2, 8))
}
