% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_factor_count.R
\name{stan_factor_count}
\alias{stan_factor_count}
\title{Count the number of levels in a factor}
\usage{
stan_factor_count(factor)
}
\arguments{
\item{factor}{A factor.}
}
\description{
This function counts the number of levels in a factor variable and is meant to be used in conjunction with \code{to_stan_factor()}. Note that both of these functions use \code{addNA()}, such that any exisiting \code{NA} values in the factor are grouped together and counted as a distinct factor level.
}
\examples{
stan_factor_count(my.factor)
}
