% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_stan_factor.R
\name{to_stan_factor}
\alias{to_stan_factor}
\title{Convert a factor variable into integers}
\usage{
to_stan_factor(factor)
}
\arguments{
\item{factor}{A factor.}
}
\description{
This function converts a factor variable into an integer vector, which is useful for coding factors in the Stan programming language. Note that \code{addNA()} is the default, so if \code{NA}s exist in the factor, they will be included in the resulting vector as their own integer value.
}
\examples{
to_stan_factor(my.factor)
}
